backend/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database and ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.0

# AutoGen AI agents
autogen-agentchat>=0.4.9

# WebSocket support
websockets>=12.0
autogen-ext[ollama]>=0.4.9
ollama>=0.1.0

# HTTP client and utilities
httpx>=0.25.0
python-multipart>=0.0.6
websockets>=12.0